+++ script/lib/utils.js
